src/main/java/com/thoughtworks/bahmnidhis2integrationservice/controller/SyncController.java
package com.thoughtworks.bahmnidhis2integrationservice.controller;

import com.thoughtworks.bahmnidhis2integrationservice.service.impl.SyncServiceImpl;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.ResponseBody;

@Controller
@RequestMapping("/api")
public class SyncController {

    @Autowired
    private SyncServiceImpl syncService;

    @PutMapping(value = "/syncData")
    @ResponseBody
    public void sync(@RequestBody String body) {
        syncService.sync(body);
    }
}
